app/backend/requirements.txt
flask
flask-cors
gunicorn
ultralytics
pillow